Added parallel functionality to manyTopics
This process was sequentially running selectModel() calls so I used functions from foreach to parallelize it and offer the option to set parallel to TRUE or FALSE and to set the number of cores used. Now, for however many K topics are being tested, that many processes can be run in parallel simultaneously.